feat: SoC styles and component logics
